Add Kill Ratio Statistics
Added functions to get the amount of cases that were opened
and closed for today and to determine what the kill ratio is
The kill ratio is simply the percentage of how many of today's
cases were closed. Ideally this should be at least 100% so that
cases are being closed slightly faster than they are opened.