Refactored Daily/Weekly stats per user
Removed the print_stats function at the end of vtiger_api.py

Split up that functionality into two functions that then return
the data instead of printing it. This data can now be used in the
app_gui.py file. It comes in as A list of tuples so that it can be
ordered from most closed cases to least closed.

The next step is to get the functionality in app_gui.py to put that
data into the two table widgets.