Bug Fix - Open Cases
Fixed a bug where only the open cases from today or from the week were
being retrieved. Every time a case was closed, that number would
decrease. Now it pulls the total number of cases that have been
opened since the beginning of the day and the total number of
cases that been closed since the beginning of the day (and week).
The kill rate's make a lot more sense.

For example

8 open 12 closed
7 open 13 closed

Now becomes

8 open 12 closed
8 open 13 closed