Threading Implementation
When gathering the data from the VTiger API and putting the data into
the GUI, the GUI would freeze. Threading was implemented using QRunnable
and QThreadpool to resolve this.

This tutorial was used primarily:
https://www.learnpyqt.com/courses/concurrent-execution/multithreading-pyqt-applications-qthreadpool/

When you press the "Manual Refresh" button, self.threading_function
creates a QRunnable 'worker' thread which gathers the data from the
VTiger API in another thread. It then returns all the data as a signal
which connects automatically to self.manual_refresh_data.
self.manual_refresh_data then takes that data and fills out the GUI.

The next step for this is to implement the auto updating feature.