Add checkboxes to show/hide stat windows
Checkboxes were added to display and hide all windows
related to stats. This is useful as we now have the platform
to auto-hide/show these based on specific inputs.
For instance, on Mondays both the week and day stats are the same.
We can auto-hide the week stats if they're the same.