Auto hide redundant week/month stats
Every Monday, the today and week stats are identical.
These stats only differentiate on Tuesday.
Instead of showing both identical day and week stats, the week
widgets are hidden if they are the same. They will be shown
once a difference is detected.

Although a much rarer occurrence, sometimes the month and the week
will have identical information. This is only when the month begins
on a Sunday or Monday. Still, in this scenario the week and month
stats will be identical for the entire week. In such a case, the
month stats are now hidden until a difference is detected.